refactor(semantic): dual binding
